Stringification refactor (start)
I noticed that the stringification with options as flags wasn't working
out very well, and there was inconsistent behaviors across some
functions over it.

Therefore this refactor is meant to make the stringification consistent
across the board, as well as provide a more easy and predictable options
of stringification.

This refactor involves the entire framework, since all base code must be
reviewed and adapted accordingly, although most of the code will remain
untouched and is compatible with this refactor.