Long method declaration coding style ehancement
I was on the fence going more towards PSR-2 or not, and in the end PSR-2
is not good enough, is too flawed from how it was defined to what the
definition ended up being, and some of the definitions of PSR-2 actually
force you to create unreadable code, in my opinion.


So while my coding style matches most of PSR-2, whichever makes no sense
to me will deviate from it, such as:

- Tabs instead of spaces for indentation (was already as such to begin
with);

- Leveraging the horizontal space available: 5 to 10 lines of nothing
more than 1 to 2 words at best for method parameters or interfaces is a
complete waste of space and makes the code less readable, as more
scrolling needs to be performed back and forth;

- Multi-line method declarations put the finishing ")" and starting "{"
in separate lines, given that now return types are a thing, and if they
get to go in the same line, the method code itself would be right in the
next line, glueing the return type and code together visually, which
defeats the entire purpose of placing a "{" in a new line in normal
declarations.