New getDefaultPropertyValue method across properties classes
Created the support for a new method to implement properties defaults
for classes such as Exception, Immutable and Options.

Before NULL was taken as a special value with no meaning and which could
be coalesced into another default value, but doing so creates ambiguity
of the expected behavior in cases when NULL is explicitly given,
therefore this method support was created o solve that ambiguity in a
clean way.