Refactored Properties trait into LazyProperties
While this and the previous refactors were dangerous and painful to do
and test, these traits have a limitation: it cannot be known all the
properties an object has, since not even the object itself knows until
it evaluates them.

Therefore, I predict that in the future I will have to create a
Properties trait where this may be possible, albeit at the cost of
lazy-loading.

However, thus far these traits seem to be used in the right classes,
except perhaps Structure, and maybe (just maybe) Immutable as well,
which will may get reevaluated so in the near future.