Properties complete refactor (start)
This addresses the problem of there being potentially 4 different
implementations for the same properties system, due to the usage of
traits.

Tests show that the lightweight implementations are only a little bit
more lightweight than the more powerful ones, and only memory-wise and
not CPU-wise, therefore it makes no sense to keep the lightweight ones.

But still 4 different traits are desireable, therefore a set of managers
will be created for each major trait:
- Properties manager for all properties traits;
- Functions manager for functions trait;
- Memoization manager for memoizations trait (which will be split into
two, static and non-static).