Input modifiers refactor: handle constraints and filters separately
Handling constraints and filters always together has caused some usage
limitations and generated some confusion at times when reading code.

With this small-ish refactor, the overall way modifiers work is still
the exact same, however constraint and filter specific methods,
properties and producer interfaces were added in order to simplify and
clarify what is a constraint and what is a filter, as well simplify a
bit further the implementation of new constraints and filters.